Combine search box and reset button
Used to look like a filter reset but it only resets the team search (by
design). Used Gumby to accomplish this, but we are currently using an
empty anchor tag to access the Gumby styles (Gumby uses the anchor tag
to set the hover click etc styles) but should change it to a button
since it isn't actually a link.